+++ changelog.md
@@ -1,3 +1,12 @@
+# 0.1.1
+
+* NEW leading ones examples
+* NEW raise exception if unknown parameters are given in the scenario file
+* FIX 17/26/35/37/38/39/40/46
+* CHANGE requirement of ConfigSpace package to 0.2.1
+* CHANGE cutoff default is now None instead of 99999999999
+
+
 # 0.1.0
 
 * Moved to github instead of bitbucket